requirements.txt
asgiref==3.9.1
babel==2.17.0
Django==5.2.6
django-filter==25.2
django-money==3.5.4
django-resized==1.0.3
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
py-moneyed==3.0
pytz==2025.2
PyYAML==6.0.3
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
